refactor: 視認性を上げるためファイル構造を変更した
