--- src/components/Noise/index.css.ts
@@ -1,12 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-const noise = style({
-  mixBlendMode: 'multiply',
-  isolation: 'isolate',
-});
-
-const styles = {
-  noise,
-};
-
-export default styles;
